+++ src/get.test.js
@@ -11,16 +11,18 @@ describe('get', () => {
       }
     }
 
-    expect(get(['name'])(target)).toBe('Jane')
-    expect(get('name')(target)).toBe('Jane')
-
-    expect(get(['age'])(target)).toBe(20)
     expect(get('age')(target)).toBe(20)
+    expect(get(['age'])(target)).toBe(20)
 
-    expect(get(['address', 'city'])(target)).toBe('Melbourne')
     expect(get('address.city')(target)).toBe('Melbourne')
+    expect(get(['address', 'city'])(target)).toBe('Melbourne')
+  })
 
-    expect(get(['foo', 'bar'])(target)).toBeUndefined()
-    expect(get('foo.bar')(target)).toBeUndefined()
+  it('handles unknown properties', () => {
+    expect(get('a')({})).toBeUndefined()
+    expect(get(['a'])({})).toBeUndefined()
+
+    expect(get('a.b')({})).toBeUndefined()
+    expect(get(['a', 'b'])({})).toBeUndefined()
   })
 })